- Enhanced Random and Complex functionality with new methods and operator overloads.
- Added `shuffle` function in Random.hpp, arithmetic operations for Complex numbers, and support for cross-type operations.

+++ NumCpp/Core/NArray.hpp
@@ -665,7 +665,7 @@ public:
 
 
     /* Index Overload */
-    NArray<dtype> operator[](const int& i) const {
+    NArray<dtype> operator[](const long long int& i) const {
         auto index = get_index(i);
 
         if (_shape.get_Ndim() == 1) {
